Move lockfile logging to osquery log adapter
