+++ pkg/osquery/runtime/osqueryinstance_windows_test.go
@@ -5,9 +5,7 @@ package runtime
 
 import (
 	"strings"
-	"syscall"
 	"testing"
-	"unsafe"
 
 	"github.com/kolide/launcher/ee/agent/types"
 	typesMocks "github.com/kolide/launcher/ee/agent/types/mocks"
@@ -51,42 +49,4 @@ func TestCreateOsqueryCommandEnvVars(t *testing.T) {
 	require.True(t, systemDriveEnvVarFound, "SystemDrive env var missing from command")
 
 	k.AssertExpectations(t)
-}
-
-func createLockFile(t *testing.T, fileToLock string) {
-	lockedFileName, err := syscall.UTF16PtrFromString(fileToLock)
-	require.NoError(t, err)
-
-	handle, err := syscall.CreateFile(
-		lockedFileName,
-		syscall.GENERIC_READ,
-		syscall.FILE_SHARE_READ|syscall.FILE_SHARE_WRITE|syscall.FILE_SHARE_DELETE,
-		nil,
-		syscall.CREATE_ALWAYS,
-		syscall.FILE_ATTRIBUTE_NORMAL,
-		0,
-	)
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		syscall.Close(handle)
-	})
-
-	// See: https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-lockfileex
-	modkernel32 := syscall.NewLazyDLL("kernel32.dll")
-	procLockFileEx := modkernel32.NewProc("LockFileEx")
-	overlapped := &syscall.Overlapped{
-		HEvent: 0,
-	}
-
-	r1, _, e1 := syscall.SyscallN(
-		procLockFileEx.Addr(),
-		uintptr(handle),
-		uintptr(2), // LOCKFILE_EXCLUSIVE_LOCK
-		uintptr(0), // Reserved parameter; must be set to zero.
-		uintptr(1), // nNumberOfBytesToLockLow -- the low-order 32 bits of the length of the byte range to lock.
-		uintptr(0), // nNumberOfBytesToLockHigh -- the high-order 32 bits of the length of the byte range to lock.
-		uintptr(unsafe.Pointer(overlapped)),
-	)
-	require.False(t, r1 == 0, error(e1))
 }
