+++ pkg/osquery/runtime/runtime_helpers.go
@@ -4,17 +4,10 @@
 package runtime
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"os/exec"
 	"path/filepath"
-	"strconv"
-	"strings"
 	"syscall"
-
-	"github.com/kolide/launcher/ee/allowedcmd"
-	"github.com/shirou/gopsutil/v3/process"
 )
 
 func setpgid() *syscall.SysProcAttr {
@@ -39,42 +32,4 @@ func platformArgs() []string {
 
 func isExitOk(_ error) bool {
 	return false
-}
-
-func getProcessesHoldingFile(ctx context.Context, pathToFile string) ([]*process.Process, error) {
-	cmd, err := allowedcmd.Lsof(ctx, "-t", pathToFile)
-	if err != nil {
-		return nil, fmt.Errorf("creating lsof command: %w", err)
-	}
-
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return nil, fmt.Errorf("running lsof: %w", err)
-	}
-
-	pidStr := strings.TrimSpace(string(out))
-	if pidStr == "" {
-		return nil, errors.New("no process found using file via lsof")
-	}
-
-	pidStrs := strings.Split(strings.TrimSpace(string(out)), "\n")
-	if len(pidStrs) == 0 {
-		return nil, errors.New("no processes found using file via lsof")
-	}
-
-	processes := make([]*process.Process, 0)
-	for _, pidStr := range pidStrs {
-		pid, err := strconv.ParseInt(pidStr, 10, 32)
-		if err != nil {
-			return nil, fmt.Errorf("invalid pid %s: %w", pidStr, err)
-		}
-
-		p, err := process.NewProcess(int32(pid))
-		if err != nil {
-			return nil, fmt.Errorf("getting process for %d: %w", pid, err)
-		}
-		processes = append(processes, p)
-	}
-
-	return processes, nil
 }
